+++ v3/aie/src/kmeans_kernel.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <adf.h>
 
-void kmeans_function(input_stream<int16_t> *restrict input, output_stream<int16_t> *restrict output);
+void kmeans_function(input_stream<float> *restrict input, output_stream<float> *restrict output);
