+++ v3/data_movers/testbench/testbench_setupaie.cpp
@@ -8,7 +8,7 @@
 #include <stdlib.h>
 #include <time.h>
 
-void read_from_stream(int16_t *buffer, hls::stream<int16_t> &stream, size_t size)
+void read_from_stream(float *buffer, hls::stream<float> &stream, size_t size)
 {
     for (size_t i = 0; i < size; i++)
     {
@@ -18,17 +18,17 @@ void read_from_stream(int16_t *buffer, hls::stream<int16_t> &stream, size_t size
 
 int main(int argc, char* argv[]) 
 {
-    hls::stream<ap_int<sizeof(int16_t) * 8 * 8>> s;
+    hls::stream<ap_int<sizeof(float) * 8 * 8>> s;
     std::srand(time(nullptr));
 
     // size := clusters coordinates (x,y) + points coordinates (x,y)
-    int16_t num_clusters = 4;
-    int16_t num_points = 8;
-    int16_t input_size = num_clusters * 2 + num_points * 2;
+    int32_t num_clusters = 4;
+    int32_t num_points = 8;
+    int32_t input_size = num_clusters * 2 + num_points * 2;
 
-    std::vector<int16_t> clusters_buffer(num_clusters * 2);
-    std::vector<int16_t> points_buffer(num_points * 2);
-    std::vector<int16_t> input_buffer(input_size);
+    std::vector<float> clusters_buffer(num_clusters * 2);
+    std::vector<float> points_buffer(num_points * 2);
+    std::vector<float> input_buffer(input_size);
 
     clusters_buffer = {3, -7, -8, 5, 10, -3, -4, -6};
     points_buffer = {7, 9, -2, 0, 5, 4, -10, -8, 6, -2, -3, 7, 1, -9, 9, 3};
@@ -43,7 +43,7 @@ int main(int argc, char* argv[])
     
     if (file.is_open()) 
     {
-        ap_int<sizeof(int16_t) * 8 * 8> tmp;
+        ap_int<sizeof(float) * 8 * 8> tmp;
 
         for (size_t i = 0; i < (input_size / 8); i++)
         {
@@ -51,7 +51,7 @@ int main(int argc, char* argv[])
 
             for (size_t j = 0; j < 8; j++) 
             {
-                int16_t x = tmp.range(j * 16 + 15, j * 16);
+                float x = tmp.range(j * 16 + 15, j * 16);
                 file << x << std::endl;
                 std::cout << x << std::endl;
             }
